# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MBKM API” Manajemen User & Role",
    "version": "1.1.0",
    "description": "Admin mengelola akun dosen (Tim MBKM, GPM, Penguji Semhas, DPL, dll - satu dosen boleh punya beberapa role) dan akun mahasiswa langsung dibuat oleh admin. Klik Authorize lalu isi token dari /auth/login untuk mencoba endpoint yang butuh login."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateDosenRequest": {
        "type": "object",
        "required": [
          "nama",
          "email",
          "password",
          "nip"
        ],
        "properties": {
          "nama": {
            "type": "string",
            "example": "Dr. Budi"
          },
          "alamat": {
            "type": "string",
            "example": "Jember"
          },
          "email": {
            "type": "string",
            "example": "budi@kampus.ac.id"
          },
          "password": {
            "type": "string",
            "example": "rahasia123"
          },
          "nip": {
            "type": "string",
            "example": "198501012010011001"
          },
          "role_ids": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "example": [
              3,
              6
            ],
            "description": "opsional, bisa lebih dari satu role sekaligus"
          }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string"
          },
          "alamat": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "nip": {
            "type": "string"
          }
        }
      },
      "SetRolesRequest": {
        "type": "object",
        "required": [
          "role_ids"
        ],
        "properties": {
          "role_ids": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "example": [
              3,
              6
            ]
          }
        }
      },
      "CreateRoleRequest": {
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "koordinator_prodi"
          },
          "name": {
            "type": "string",
            "example": "Koordinator Prodi"
          },
          "description": {
            "type": "string"
          },
          "allowed_user_types": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "dosen"
            ]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Daftar role (untuk pilihan/checkbox)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "dosen",
                "mahasiswa"
              ]
            },
            "description": "filter: role yang boleh dipegang tipe akun ini"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "Buat role baru (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dibuat"
          }
        }
      }
    },
    "/roles/{id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Detail role",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "Edit role (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "summary": "Hapus role (admin, role bawaan terlindungi)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users (Dosen)"
        ],
        "summary": "List akun dosen (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "filter kode role, mis. tim_mbkm"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "cari nama/email/NIP"
          },
          {
            "name": "is_active",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Users (Dosen)"
        ],
        "summary": "Admin membuat akun dosen baru (boleh langsung isi role_ids)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDosenRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dibuat"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users (Dosen)"
        ],
        "summary": "Detail akun dosen",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Users (Dosen)"
        ],
        "summary": "Edit akun dosen (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Users (Dosen)"
        ],
        "summary": "Hapus akun dosen (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/{id}/roles": {
      "get": {
        "tags": [
          "Users (Dosen)"
        ],
        "summary": "Role yang dipegang dosen ini",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Users (Dosen)"
        ],
        "summary": "Ganti SEMUA role dosen sekaligus (checkbox form)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRolesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Users (Dosen)"
        ],
        "summary": "Tambah role ke dosen tanpa menghapus yang lama",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRolesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/{id}/roles/{roleId}": {
      "delete": {
        "tags": [
          "Users (Dosen)"
        ],
        "summary": "Cabut satu role dari dosen",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}